+++ LeagueAPI/LeagueAPI/Models/RiotModels/BannedChampion.swift
@@ -10,8 +10,8 @@ import Foundation
 
 public class BannedChampion: Decodable {
     
-    public var championId: Double
-    public var teamId: Double
+    public var championId: Int64
+    public var teamId: Int64
     public var banTurn: Int
     
     enum CodingKeys: String, CodingKey {
@@ -20,7 +20,7 @@ public class BannedChampion: Decodable {
         case banTurn = "pickTurn"
     }
     
-    public init(championId: Double, teamId: Double, banTurn: Int) {
+    public init(championId: Int64, teamId: Int64, banTurn: Int) {
         self.championId = championId
         self.teamId = teamId
         self.banTurn = banTurn
@@ -28,8 +28,8 @@ public class BannedChampion: Decodable {
     
     public required init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.championId = try container.decode(Double.self, forKey: .championId)
-        self.teamId = try container.decode(Double.self, forKey: .teamId)
+        self.championId = try container.decode(Int64.self, forKey: .championId)
+        self.teamId = try container.decode(Int64.self, forKey: .teamId)
         self.banTurn = try container.decode(Int.self, forKey: .banTurn)
     }
 }
