+++ LeagueAPI/LeagueAPI/Models/RiotModels/Participant.swift
@@ -11,13 +11,13 @@ import Foundation
 public class Participant: Decodable {
     
     public var summonerName: String
-    public var profileIconId: Double
-    public var championId: Double
+    public var profileIconId: Int64
+    public var championId: Int64
     public var isBot: Bool
-    public var teamId: Double
-    public var summonerSpell1: Double
-    public var summonerSpell2: Double
-    public var summonerId: Double? // Detailed participant field
+    public var teamId: Int64
+    public var summonerSpell1: Int64
+    public var summonerSpell2: Int64
+    public var summonerId: Int64? // Detailed participant field
     public var runePage: RunePage? // Detailed participant field
     public var customizedObjects: [GameCustomObject]? // Detailed participant field
     
@@ -34,7 +34,7 @@ public class Participant: Decodable {
         case customizedObjects = "gameCustomizationObjects"
     }
     
-    public init(summonerName: String, profileIconId: Double, championId: Double, isBot: Bool, teamId: Double, summonerSpell1: Double, summonerSpell2: Double, summonerId: Double? = nil, runePage: RunePage? = nil, customizedObjects: [GameCustomObject]? = nil) {
+    public init(summonerName: String, profileIconId: Int64, championId: Int64, isBot: Bool, teamId: Int64, summonerSpell1: Int64, summonerSpell2: Int64, summonerId: Int64? = nil, runePage: RunePage? = nil, customizedObjects: [GameCustomObject]? = nil) {
         self.summonerName = summonerName
         self.profileIconId = profileIconId
         self.championId = championId
@@ -50,13 +50,13 @@ public class Participant: Decodable {
     public required init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         self.summonerName = try container.decode(String.self, forKey: .summonerName)
-        self.profileIconId = try container.decode(Double.self, forKey: .profileIconId)
-        self.championId = try container.decode(Double.self, forKey: .championId)
+        self.profileIconId = try container.decode(Int64.self, forKey: .profileIconId)
+        self.championId = try container.decode(Int64.self, forKey: .championId)
         self.isBot = try container.decode(Bool.self, forKey: .isBot)
-        self.teamId = try container.decode(Double.self, forKey: .teamId)
-        self.summonerSpell1 = try container.decode(Double.self, forKey: .summonerSpell1)
-        self.summonerSpell2 = try container.decode(Double.self, forKey: .summonerSpell2)
-        self.summonerId = try container.decode(Double?.self, forKey: .summonerId)
+        self.teamId = try container.decode(Int64.self, forKey: .teamId)
+        self.summonerSpell1 = try container.decode(Int64.self, forKey: .summonerSpell1)
+        self.summonerSpell2 = try container.decode(Int64.self, forKey: .summonerSpell2)
+        self.summonerId = try container.decode(Int64?.self, forKey: .summonerId)
         self.runePage = try container.decode(RunePage?.self, forKey: .runePage)
         self.customizedObjects = try container.decode([GameCustomObject]?.self, forKey: .customizedObjects)
     }
